Corrected Task 16

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,18 +6,18 @@
  */
 int main(void)
 {
-	unsigned long int num1, num2, temp;
+	long double num1, num2, temp;
 	short i;
 
 	num1 = 1;
 	num2 = 2;
-	printf("%lu, %lu, ", num1, num2);
+	printf("%.0Lf, %.0Lf, ", num1, num2);
 	for (i = 0; i < 96; i++)
 	{
 		temp = num2;
 		num2 = num2 + num1;
 		num1 = num2;
-		printf("%lu, ", temp);
+		printf("%.0Lf, ", temp);
 	}
 	printf("\n");
 	return (0);
